Se corrige editar producto

+++ src/controllers/product-controller.js
@@ -140,11 +140,14 @@ const controller = {
     let colors=req.body.colors
     let image={}
     let images_id={}
+    console.log(product)
     let productToUpdate=await Product.update(product,{where:{
       id:req.params.id,
     }});
-    if(req.files.length!=0){
-      console.log(req.files)
+
+  console.log(req.files)
+    if(req.files.length!==0){
+    
     for (let i = 0; i < 5; i++) {
        image={name_archive:req.files[i] ? req.files[i].filename : "default-image.png",
               product_id:req.params.id           
@@ -158,6 +161,8 @@ const controller = {
     }
   }
    //Guarda en la tabla intermedia ProductSizes
+if (sizes) {
+  
 
      await ProductSizes.destroy({where:{
   product_id:req.params.id
@@ -168,8 +173,10 @@ const controller = {
     product_id:req.params.id,
     size_id:sizes[i]
      }) }
-
+    }
 //Guarda en la tabla intermedia ProductColors
+if (colors) {
+  
 
 await ProductColors.destroy({where:{
   product_id:req.params.id
@@ -179,7 +186,7 @@ await ProductColors.destroy({where:{
      await ProductColors.create({
     product_id:req.params.id,
     color_id:colors[i]
-     }) }
+     }) }}
 
      res.redirect("/dashboard/product"); 
   },
